+++ src/main/java/com/time/studentmanage/StudentmanageApplication.java
@@ -1,5 +1,9 @@
 package com.time.studentmanage;
 
+import com.time.studentmanage.config.TestDataInit;
+import com.time.studentmanage.repository.ParentRepository;
+import com.time.studentmanage.repository.StudentRepository;
+import com.time.studentmanage.repository.TeacherRepository;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
