+++ src/main/java/com/time/studentmanage/repository/ParentRepository.java
@@ -1,4 +1,4 @@
-package com.time.studentmanage;
+package com.time.studentmanage.repository;
 
 import com.time.studentmanage.domain.member.Parent;
 import org.springframework.data.jpa.repository.JpaRepository;
